fix(extensions): moved link enhanceto markdown preset
